Replaces reuse of Query Pagination block in part with duplication in template.
The Query Pagination (QP) block was in a template part so it could be reused, resulting in an extra nested layer. However, due to a bug in WP, it can only be used in templates. This reduces the nesting, but results in duplication of code instead of reuse.

Details of the bug: https://github.com/WordPress/gutenberg/issues/38684