Improves and standardises formatting of comments.
Formats metadata the same as article metadata is formatted.

Numbers the comments with nested Arabic numbers.

Improves the wording of the previous/next links, which are reused for the comment pagination.

Extend WP's Walker class to generate the HTML for the comments so the default doesn't have to be used.