Adds header with cover image.
The cover image is a WP Cover block, which uses an image in assets. The image can only be included using PHP, so the Cover block itself is included via a pattern.

If a pattern has the same name as a template part, the pattern will be invoked, not the part.

It seems the Cover block must be last: anything following it does not appear. The header with cover image uses the grid display to position the navigation menu after the Cover block.

The header with cover image is used in the focus page and focus post. The footer has been change din each to the fancy footer.