Makes footer sticky.
Uses MDN guide: https://developer.mozilla.org/en-US/docs/Web/CSS/Layout_cookbook/Sticky_footers. The height of the grid wrapper (.wp-site-blocks) must exclude the grid row gap.

Other adjustments prevent the grid row heights from stretching inconsistently.

The comment form is given a class so that the form can tell whether the user is logged in or not, and hence use the correct number of rows so that unnecessary row gaps are not added.