Adds borders to various UI elements.
Thereseems to be a bug in Safari that causes it to display the comment textarea (for input of the comment body) in the wrong colour unless the width is set to >=2px or the style is set to dotted or dashed.

Firefox, Chrome and Opera display the border as expected.