Uses alternative method to create inset/outset borders for UI elements.
Inset and outset borders aren't necessarily very obvious in Firefox. This takes a base colour and creates noticeably lighter and darker colours from it, and uses those as the border's colours to give the appearance of inset and outset borders. The problem is that the base colour must be determined before runtime, which is not good if the colour is changed elsewhere.