Removes .is-outline-mode class from the header.
It doesn't work. When a (full) template is viewed in the Site Editor (SE), the header works as it's meant to. When a part (template) is viewed in the SE, it no longer overlaps, but it is not laid out according to the specified CSS grid (the layout is obtained from the HTML order instead). In the SE, the parent component is always the same, no matter what (template) part is being looked at, so it's impossible to specify a layout that depends on the class of its parent.